requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.8
django-crispy-forms==1.8.1
idna==2.8
oauthlib==3.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
django-qr-code==1.1.0
djangorestframework
mysqlclient==1.4.4
pillow==6.2.1
django-environ==0.4.5
django-extensions==2.2.5